ble_heartrate_reader/requirements.txt
click==8.0.3
Flask==2.0.2
importlib-metadata==4.11.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.28
numpy==1.21.5
pkg-resources==0.0.0
protobuf==3.19.4
typing-extensions==4.0.1
Werkzeug==2.0.3
zipp==3.7.0
